Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/aspect-ratio-affects-container-width-when-height-changes-expected.txt

There should be a green square below, and no red.



PASS Changing height should affect the descendant widths, due to aspect ratio 
FAIL #container 1 assert_equals: 
<div id="container" style="height: 100px;">
  <div style="height:100%;" data-expected-height="100">
    <div style="float:left; height:100%; background:red;" data-expected-width="100" data-expected-height="100">
      <!-- The image is a 1x1 transparent one. -->
      <img style="display:block; height:100%; background:green;" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7" data-expected-width="100" data-expected-height="100">
    </div>
  </div>
</div>
width expected 100 but got 0
